requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
atomicwrites==1.4.1
attrs==25.4.0
av==16.0.1
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
ctranslate2==4.6.2
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.17
decorator==4.4.2
deepface==0.0.96
defusedxml==0.7.1
exceptiongroup==1.3.1
executing==2.2.1
facenet-pytorch==2.6.0
faster-whisper==1.2.1
fastjsonschema==2.21.2
fer==22.5.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.61.0
fqdn==1.5.1
frozendict==2.4.7
fsspec==2025.9.0
future==1.0.0
gast==0.7.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
gpt4all==2.8.2
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.21.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keras==3.12.0
kiwisolver==1.4.9
lark==1.3.1
lazr.uri==1.0.7
libclang==18.1.1
lightecc==0.0.3
lightphe==0.0.19
llvmlite==0.45.1
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
moviepy==1.0.3
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.0
multitasking==0.0.12
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.5.0
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
peewee==3.18.3
pillow==10.2.0
platformdirs==4.5.1
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
proglog==0.1.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
py==1.11.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyllamacpp==2.4.3
pynndescent==0.5.13
pyparsing==3.2.5
pyreadline3==3.5.4
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyWavelets==1.8.0
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==1.8.3
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sounddevice==0.5.3
soupsieve==2.8
srsly==2.5.2
stack-data==0.6.3
statsmodels==0.14.6
stumpy==1.13.0
sympy==1.13.1
tcmlib==1.4.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
terminado==0.18.1
tf_keras==2.20.1
threadpoolctl==3.6.0
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tornado==6.5.1
tqdm==4.67.1
traitlets==5.7.1
traittypes==0.2.3
transformers==4.57.2
treelite==4.4.1
treescope==0.1.10
tsfresh==0.21.1
tweepy==4.16.0
typeguard==4.4.4
typer==0.20.0
typer-slim==0.20.0
types-pytz==2025.2.0.20251108
types-setuptools==80.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
ultralytics==8.3.235
ultralytics-thop==2.0.18
umap-learn==0.5.9.post2
umf==1.0.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vega-datasets==0.9.0
wadllib==1.3.6
wandb==0.23.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==3.6.10
wordcloud==1.9.4
wrapt==2.0.1
wurlitzer==3.1.1
xarray==2025.6.1
xarray-einstats==0.8.0
xgboost==3.1.2
xlrd==2.0.2
xxhash==3.6.0
xyzservices==2025.11.0
yarl==1.22.0
ydf==0.8.0
yellowbrick==1.5
yfinance==0.2.66
zict==3.0.0
zipp==3.23.0
zstandard==0.25.0
